fix: :adhesive_bandage: Type: Refactoring | Title: Minor Code Refactoring and Configuration Update
Description:
The changes mainly involve minor code refactoring and configuration updates. The refactoring includes the correction of a typo in a confirmation message and the addition of a new remote pattern in the Next.js configuration. The configuration update allows the application to handle images from the 'cloud.appwrite.io' hostname.

Main Files Walkthrough:
- README.md: No actual changes were made in this file. The changes seen in the diff are due to the conversion of line endings from Unix-style (LF) to Windows-style (CRLF).
- app/post/_components/single-comment.tsx: Corrected a typo in the confirmation message displayed when deleting a comment. The word "weant" was changed to "want".
- hooks/use-change-user-image.ts: Corrected a typo in the environment variable name. The word "DEAFULT" was changed to "DEFAULT".  next.config.js: Added a new remote pattern to the Next.js image configuration. This allows the application to handle images from the 'cloud.appwrite.io' hostname.

Fix on images and typos.