refactor: :children_crossing: Type: Improvements ,Refactoring | Title: Refactoring HomePage Component to Use Post Store
Description:
The HomePage component has been refactored to use the post store instead of hard-coded data. The useEffect hook is used to set all posts when the component mounts. The posts are then mapped to the PostMain component for display.

Main Files Walkthrough:
- app/(main)/(home)/page.tsx: The HomePage component now imports the usePostStore hook from the post store. A useEffect hook has been added to call the setAllPosts function from the post store when the component mounts. The hard-coded post data has been replaced with a map function that renders a PostMain component for each post in the allPosts array from the post store.

Improved home page to receive dynamic data and render it.