feat: :children_crossing: Type: Improvements, Enhancement, Refactoring | Title: Improvements on ProfilePage, PostPag, TopNav and EditProfileOverlay
Description:
The changes involve the removal of the ClientOnly component in the TopNav component. This component was previously used to render the search profiles dropdown only on the client side. However, it has been removed, and the dropdown is now rendered directly without the ClientOnly wrapper. The modified components: EditProfileOverlay ProfilePage TopNav PostPage.

Main Files Walkthrough:
- app/_components/top-nav.tsx: The ClientOnly component was removed from the search profiles dropdown rendering. The dropdown is now rendered directly without the ClientOnly wrapper.
- EditProfileOverlay app/(main)/profile/[id]/_components/edit-profile-overlay.tsx.
- ProfilePage app/(main)/profile/[id]/page.tsx.
- TopNav app/_components/top-nav.tsx.
- PostPage app/post/[postId]/[userId]/page.tsx.

Final improvements on overall functionality and performance.