fix: :arrow_down: fix(package): Update @types/node version in package-lock.json
Description:

This commit updates the version of @types/node in package-lock.json from 20.13.0 to 20.14.0. This change is necessary to keep the project dependencies up-to-date and ensure compatibility with the latest features and improvements provided by the Node.js type definitions.

Details:

The @types/node package provides TypeScript definitions for Node.js, enabling TypeScript to understand the types used in Node.js. By updating this package, we ensure that our project can leverage the latest type definitions provided by Node.js, which can include new features, bug fixes, and performance improvements.

Changeset:

The change was made in the package-lock.json file. The version of @types/node was updated from 20.13.0 to 20.14.0.

Modified files:

1. package-lock.json: The version of @types/node was updated from 20.13.0 to 20.14.0.

This commit ensures that the project stays up-to-date with the latest Node.js type definitions, which can lead to better type checking, improved IntelliSense, and access to new features.

Signed-off-by: RicardoGEsteves(Ricardo Esteves) 30535937+RicardoGEsteves@users.noreply.github.com

Refs: [#162]
Related to: [#162]